add repository

+++ src/main/java/com/codegym/controller/ProductController.java
@@ -13,7 +13,6 @@ import org.springframework.web.servlet.ModelAndView;
 
 import java.io.File;
 import java.io.IOException;
-import java.util.List;
 import java.util.Optional;
 
 @Controller
@@ -27,7 +26,7 @@ public class ProductController {
     @GetMapping("/products/list")
     public ModelAndView showListProduct(@RequestParam(name = "q") Optional<String> q) {
         ModelAndView modelAndView = new ModelAndView("/product/list");
-        List<Product> products = productService.findAll();
+        Iterable<Product> products = productService.findAll();
         if (q.isPresent()) {
             products = productService.findProductByNameContaining(q.get());
         }
@@ -37,21 +36,27 @@ public class ProductController {
 
     @GetMapping("/products/delete/{id}")
     public ModelAndView showDeleteProduct(@PathVariable Long id) {
+        Optional<Product> productOptional = productService.findById(id);
+        if (!productOptional.isPresent()) {
+            return new ModelAndView("error-404");
+        }
         ModelAndView modelAndView = new ModelAndView("/product/delete");
-        Product product = productService.findById(id);
-        modelAndView.addObject("product", product);
+        modelAndView.addObject("product", productOptional.get());
         return modelAndView;
     }
 
     @PostMapping("/products/delete/{id}")
     public ModelAndView deleteProduct(@PathVariable Long id) {
-        Product product = productService.findById(id);
-        File file = new File(uploadPath + product.getImage());
-        if (file.exists()) {
-            file.delete();
+        Optional<Product> product = productService.findById(id);
+        if (product.isPresent()) {
+            File file = new File(uploadPath + product.get().getImage());
+            if (file.exists()) {
+                file.delete();
+            }
+            productService.removeById(id);
+            return new ModelAndView("redirect:/products/list");
         }
-        productService.removeById(id);
-        return new ModelAndView("redirect:/products/list");
+        return new ModelAndView("error-404");
     }
 
     @GetMapping("/products/create")
@@ -78,37 +83,57 @@ public class ProductController {
 
     @GetMapping("/products/edit/{id}")
     public ModelAndView showEditForm(@PathVariable Long id) {
+        Optional<Product> productOptional = productService.findById(id);
+        if (!productOptional.isPresent()) {
+            return new ModelAndView("error-404");
+        }
         ModelAndView modelAndView = new ModelAndView("/product/edit");
-        Product product = productService.findById(id);
-        modelAndView.addObject("product", product);
+        modelAndView.addObject("product", productOptional.get());
         return modelAndView;
     }
 
     @PostMapping("/products/edit/{id}")
     public ModelAndView editProduct(@PathVariable Long id, @ModelAttribute ProductForm productForm) {
         MultipartFile img = productForm.getImage();
-        Product oldProduct = productService.findById(id);
-        if (img.getSize() != 0) {
-            String fileName = productForm.getImage().getOriginalFilename();
-            long currentTime = System.currentTimeMillis(); //Xử lý lấy thời gian hiện tại
-            fileName = currentTime + fileName;
-            oldProduct.setImage(fileName);
-            try {
-                FileCopyUtils.copy(img.getBytes(), new File(uploadPath + fileName));
-            } catch (IOException e) {
-                e.printStackTrace();
+        Optional<Product> product = productService.findById(id);
+        if (product.isPresent()) {
+            Product oldProduct = product.get();
+            if (img.getSize() != 0) {
+                String fileName = productForm.getImage().getOriginalFilename();
+                long currentTime = System.currentTimeMillis(); //Xử lý lấy thời gian hiện tại
+                fileName = currentTime + fileName;
+                oldProduct.setImage(fileName);
+                try {
+                    FileCopyUtils.copy(img.getBytes(), new File(uploadPath + fileName));
+                } catch (IOException e) {
+                    e.printStackTrace();
+                }
             }
+            oldProduct.setPrice(productForm.getPrice());
+            oldProduct.setDescription(productForm.getDescription());
+            oldProduct.setName(productForm.getName());
+            productService.save(oldProduct);
+            return new ModelAndView("redirect:/products/list");
         }
-        oldProduct.setPrice(productForm.getPrice());
-        oldProduct.setDescription(productForm.getDescription());
-        oldProduct.setName(productForm.getName());
-        productService.save(oldProduct);
-        return new ModelAndView("redirect:/products/list");
+        return new ModelAndView("error-404");
     }
 
     @GetMapping("/products/{id}")
     public ModelAndView showProductDetail(@PathVariable Long id) {
-        Product product = productService.findById(id);
-        return new ModelAndView("/product/view", "product", product);
+        Optional<Product> productOptional = productService.findById(id);
+        if (!productOptional.isPresent()) {
+            return new ModelAndView("error-404");
+        }
+        ModelAndView modelAndView = new ModelAndView("/product/view");
+        modelAndView.addObject("product", productOptional.get());
+        return modelAndView;
+    }
+
+    @GetMapping("/products/search")
+    public ModelAndView showProductSearch(@RequestParam("min") Double min, @RequestParam("max") Double max) {
+        Iterable<Product> products = productService.findProductPriceBetween(min, max);
+        ModelAndView modelAndView = new ModelAndView("/product/list");
+        modelAndView.addObject("products", products);
+        return modelAndView;
     }
 }
